Update README.md
The Python code written here uses Tensor Flow to create a model that can feature saVVY advanced aerospace concept design model. The code first breaks down user inputs into individual features that are then converted into input vectors, which are processed by a Savvy Chapter API via a tf.keras.Sequential model. Autopilot is used to automate the pre-processing, normalizing, batching, and training of the model. The model is validated and iterated on, and linked to a main CNN model. A 2D sketch feature is added to the model, followed by updating the model and compiling it. The model is evaluated and used to design advanced vehicles, and its performance is monitored and updated throughout the process. Finally, the saVVy Tensor Flow Model is finalized once the model is trained and optimized.